Move all background related code to background folder
